requirements.txt
google-cloud-monitoring==2.6.0